Encapsulation
Ref #14
Ref #11
