--- YOUR_ADMIN/includes/extra_datafiles/xsell_advanced_definitions.php
@@ -1,16 +1,0 @@
-<?php
-/**
- * Cross Sell Advanced
- *
- * Derived from:
- * Original Idea From Isaac Mualem im@imwebdesigning.com <mailto:im@imwebdesigning.com>
- * Portions Copyright (c) 2002 osCommerce
- * Complete Recoding From Stephen Walker admin@snjcomputers.com
- * Released under the GNU General Public License
- *
- * Adapted to Zen Cart by Merlin - Spring 2005
- * Reworked for Zen Cart v1.3.0  03-30-2006
- * 
- * This file is obsolete in Cross Sell Advance v1.3. You can remove this file from your store! 
- * 
- */
