--- YOUR_ADMIN/includes/functions/extra_functions/zen_cfg_read_only.php
@@ -1,19 +1,0 @@
-<?php
-/**
- * zen_cfg_read_only.php
- *
- * @package
- * @copyright Copyright 2004-2015 Andrew Berezin eCommerce-Service.com
- * @copyright Copyright 2003-2015 Zen Cart Development Team
- * @license http://www.zen-cart.com/license/2_0.txt GNU Public License V2.0
- * @version $Id: zen_cfg_read_only.php, v 1.1 22.01.2015 12:52:38 AndrewBerezin $
- */
-if (!function_exists('zen_cfg_read_only')) {
-    function zen_cfg_read_only($text, $key = '')
-    {
-        $name = (($key) ? 'configuration[' . $key . ']' : 'configuration_value');
-        $text = htmlspecialchars_decode($text);
-
-        return $text . zen_draw_hidden_field($name, $text);
-    }
-}
