--- YOUR_ADMIN/includes/languages/english/extra_definitions/lang.x-sell_defs.php
@@ -1,31 +1,0 @@
-<?php
-/**
- * Cross Sell Advanced
- *
- * Derived from:
- * Original Idea From Isaac Mualem im@imwebdesigning.com <mailto:im@imwebdesigning.com>
- * Portions Copyright (c) 2002 osCommerce
- * Complete Recoding From Stephen Walker admin@snjcomputers.com
- * Released under the GNU General Public License
- *
- * Adapted to Zen Cart by Merlin - Spring 2005
- * Reworked for Zen Cart v1.3.0  03-30-2006
- * Reworked for Zen Cart 1.5.7+, December 2021, lat9.
- * Reworked for Zen Cart 1.5.8+, lat9, May 2026
- */
-
-//XSell Admin Menu Boxes
-$define = [
-    'BOX_CONFIGURATION_XSELL' => 'Cross-Sell Advanced II',
-    'BOX_CATALOG_XSELL' => 'Cross-Sell Advanced II',
-    'BOX_CATALOG_XSELL_ADVANCED' => 'Cross Sell (X-Sell) Advanced Admin',
-
-    // -----
-    // Install/upgrade messages, used by init_xsell_config.php.
-    //
-    'MESSAGE_XSELL_INSTALLED' => 'Cross Sell Advanced II (v%s) was successfully installed.',
-    'MESSAGE_XSELL_UPDATED' => 'Cross Sell Advanced II was updated to v%s',
-    'MESSAGE_XSELL_DUPLICATES_REMOVED' => '%u duplicate cross-sell entries were removed.',
-    'MESSAGE_XSELL_REMOVED' => '%u cross-sell entries were removed; the associated products no longer exist.',
-];
-return $define;
