--- includes/languages/english/extra_definitions/lang.xsell_products_definitions.php
@@ -1,7 +1,0 @@
-<?php
-// -----
-// Part of the "Cross Sell Advanced" plugin for Zen Cart 1.5.8 and later.
-//
-return [
-    'TEXT_XSELL_PRODUCTS' => 'Related Products',
-];
