+++ zc_plugins/ProductsXsell/v3.0.0/Installer/languages/english/main.php
@@ -1,0 +1,9 @@
+<?php
+// -----
+// Part of the installation script for the "encapsulated" Cross-Sell Advanced II plugin for Zen Cart, by lat9.
+//
+zen_define_default('ERROR_REMOVE_FILES_CANT_DELETE', 'Unable to remove file: %s');
+zen_define_default('ERROR_REMOVE_FILES_CONTEXT', 'Invalid context supplied (%s), it must be either "catalog" or "admin".');
+
+zen_define_default('MESSAGE_XSELL_DUPLICATES_REMOVED', '%u duplicate cross-sell entries were removed.');
+zen_define_default('MESSAGE_XSELL_REMOVED', '%u cross-sell entries were removed; the associated products no longer exist.');
