+++ zc_plugins/ProductsXsell/v3.0.0/admin/includes/extra_datafiles/xsell_definitions.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Cross Sell Advanced
+ *
+ * Derived from:
+ * Original Idea From Isaac Mualem im@imwebdesigning.com
+ * Portions Copyright (c) 2002 osCommerce
+ * Complete Recoding From Stephen Walker admin@snjcomputers.com
+ * Released under the GNU General Public License
+ *
+ * Adapted to Zen Cart by Merlin - Spring 2005
+ * Reworked for Zen Cart v1.3.0  03-30-2006
+ */
+
+// XSell database
+define('TABLE_PRODUCTS_XSELL', DB_PREFIX . 'products_xsell');
+
+//XSell Catalog Menu filenames
+define('FILENAME_XSELL', 'xsell');
