+++ zc_plugins/ProductsXsell/v3.0.0/admin/includes/languages/english/extra_definitions/lang.x-sell_defs.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * Cross Sell Advanced
+ *
+ * Derived from:
+ * Original Idea From Isaac Mualem im@imwebdesigning.com <mailto:im@imwebdesigning.com>
+ * Portions Copyright (c) 2002 osCommerce
+ * Complete Recoding From Stephen Walker admin@snjcomputers.com
+ * Released under the GNU General Public License
+ *
+ * Adapted to Zen Cart by Merlin - Spring 2005
+ * Reworked for Zen Cart v1.3.0  03-30-2006
+ * Reworked for Zen Cart 1.5.7+, December 2021, lat9.
+ * Reworked for Zen Cart 1.5.8+, lat9, May 2026
+ */
+$define = [
+    'BOX_CONFIGURATION_XSELL' => 'Cross-Sell Advanced II',
+    'BOX_CATALOG_XSELL' => 'Cross-Sell Advanced II',
+];
+return $define;
