+++ zc_plugins/ProductsXsell/v3.0.0/catalog/includes/extra_datafiles/xsell_definitions.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * Cross Sell Advanced II
+ *
+ * Derived from:
+ * Original Idea From Isaac Mualem im@imwebdesigning.com
+ * Portions Copyright (c) 2002 osCommerce
+ * Complete Recoding From Stephen Walker admin@snjcomputers.com
+ * Released under the GNU General Public License
+ *
+ * Adapted to Zen Cart by Merlin - Spring 2005
+ * Reworked for Zen Cart v1.3.0  03-30-2006
+ * Reworked for Zen Cart 1.5.7+, lat9, December 2021
+ */
+
+// database
+define('TABLE_PRODUCTS_XSELL', DB_PREFIX . 'products_xsell');
+
+//filenames
+define('FILENAME_XSELL2_PRODUCTS', 'xsell2_products.php');
