+++ zc_plugins/ProductsXsell/v3.0.0/catalog/includes/languages/english/extra_definitions/lang.xsell_products_definitions.php
@@ -1,0 +1,7 @@
+<?php
+// -----
+// Part of the "Cross Sell Advanced" plugin for Zen Cart 2.1.0 and later.
+//
+return [
+    'TEXT_XSELL_PRODUCTS' => 'Related Products',
+];
