+++ zc_plugins/ProductsXsell/v3.0.0/manifest.php
@@ -1,0 +1,13 @@
+<?php
+return [
+    'pluginVersion' => 'v3.0.0',
+    'pluginName' => 'Cross-Sell Advanced II',
+    'pluginDescription' => 'This plugin enables a site to &quot;cross-sell&quot; products.  An admin tool is provided that is used to \'connect\' the various cross-sells for the site and a simple storefront \'centerbox\' is provided for the cross-sells\' display. See the plugin\'s <a href="https://github.com/lat9/cross_sell_advanced/wiki" target="_blank" rel="noopener noreferrer">Wiki</a> for additional information.',
+    'pluginAuthor' => 'Vinos de Frutas Tropicales (lat9)',
+    'pluginId' => 2334, // ID from Zen Cart forum
+    'zcVersions' => [],
+    'changelog' => '', // online URL (eg github release tag page, or changelog file there) or local filename only, ie: changelog.txt (in same dir as this manifest file)
+    'github_repo' => 'https://github.com/lat9/cross_sell_advanced', // url
+    'pluginGroups' => [],
+    'removesUnencapsulatedVersion' => true, // !empty presence indicates that non-encapsulated files are removed during install
+];
